[FEAT] NicknameSettingView 구현 및 컴포넌트 리팩토링
